requirements.txt
google-genai
google-api-python-client
google-auth
python-dotenv
pydantic